+++ tests/harness.lua
@@ -169,6 +169,7 @@ load("Spec_Elemental.lua")
 load("Spec_Devourer.lua")
 load("Spec_BeastMastery.lua")
 load("Spec_Marksmanship.lua")
+load("Spec_Retribution.lua")
 
 PRIO.Engine:OnSpecChanged()
 H.spec = PRIO.specs[269]
@@ -178,6 +179,7 @@ H.eleSpec = PRIO.specs[262]
 H.devourerSpec = PRIO.specs[1480]
 H.bmSpec = PRIO.specs[253]
 H.mmSpec = PRIO.specs[254]
+H.retSpec = PRIO.specs[70]
 H.Engine = PRIO.Engine
 H.Cond = PRIO.Cond
 H.API = API